ddl: migrate test-infra to testify for issue tests
